pyproject.toml
[build-system]
requires = ["maturin>=1.4,<2.0", "cmake>=3.15"]
build-backend = "maturin"

[project]
name = "argweavers"
description = "Ancestral recombination graph sampling method"
license = { file = "LICENSE" }
authors = [
    {name = "Matt Rasmussen", email = "matt.rasmus@gmail.com"},
    { name = "Melissa Hubisz", email = "mjhubisz@gmail.com" },
]
maintainers = [{name = "Tang Ziya", email = "tcztzy@gmail.com"}]
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: C++",
    "Programming Language :: R",
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
dynamic = ["version"]
dependencies = [
    "scipy>=1.10.1",
    "numpy>=1.24.4",
    "polars>=0.20.9",
    "pyarrow>=15.0.0",
    "rpy2>=3.5.15",
    "pandas>=2.0.3",
    "matplotlib>=3.7.5",
    "biopython>=1.83",
]

[project.scripts]
arg-likelihood = "argweavers.bin:arg_likelihood"
arg-sample = "argweavers.bin:arg_sample"
arg-summarize = "argweavers.bin:arg_summarize"
smc2bed = "argweavers.scripts.smc2bed:main"
smc2bed-all = "argweavers.scripts.smc2bed_all:main"

[tool.rye]
dev-dependencies = [
    "pip>=23.3.2",
    "pre-commit>=3.5.0",
    "pytest>=8.0",
    "pytest-cov>=4.1.0",
    "sphinx>=7.1.2",
    "esbonio-extensions>=0.2.2",
    "myst-parser>=2.0.0",
    "sphinx-autodoc2>=0.5.0",
    "ipykernel>=6.29.2",
    "cibuildwheel>=2.16.5",
]
[tool.rye.scripts]
cov-report = { chain = ["htmlcov", "htmlcov-serve"] }
htmlcov = "pytest --cov-report=html"
htmlcov-serve = "python -m http.server --directory htmlcov"

[tool.maturin]
features = ["extension-module"]
include = ["bin/*"]

[[tool.mypy.overrides]]
module = [
    "rpy2.*",
]
ignore_missing_imports = true

[tool.ruff]
extend-select = ["I"]
target-version = "py38"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = ["--cov=argweavers"]

[tool.coverage.report]
exclude_also = [
    "if (typing\\.)?TYPE_CHECKING",
    "if __name__ == \"__main__\"",
    "class \\w+\\((argparse\\.)?HelpFormatter\\)",
]

[tool.cibuildwheel]
skip = "pp*"
